refac(google): make code easier to maintain
